# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/workflow/workflow/{id}": {
      "get": {
        "operationId": "WorkflowController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for get ToolsWorkflowDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow"
        ]
      },
      "put": {
        "operationId": "WorkflowController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolsWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response for put ToolsWorkflowDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow"
        ]
      },
      "delete": {
        "operationId": "WorkflowController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for delete ToolsWorkflowDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow"
        ]
      }
    },
...
    "/auth/workflow_step/workflow_step/{id}": {
      "get": {
        "operationId": "WorkflowStepController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for get ToolsWorkflowStepDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowStepDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow_step"
        ]
      },
      "put": {
        "operationId": "WorkflowStepController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolsWorkflowStepDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response for put ToolsWorkflowStepDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowStepDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow_step"
        ]
      },
      "delete": {
        "operationId": "WorkflowStepController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for delete ToolsWorkflowStepDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowStepDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow_step"
        ]
      }
    },
    "/auth/workflow_step/workflow_step_report/list": {
      "post": {
        "operationId": "WorkflowStepReportController_findAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaseQueryDtoSmart"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response for list ToolsWorkflowStepReportDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ToolsWorkflowStepReportDto"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 123
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow_step_report"
        ]
      }
    },
    "/auth/workflow_step/workflow_step_report/{id}": {
      "get": {
        "operationId": "WorkflowStepReportController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for get ToolsWorkflowStepReportDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowStepReportDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow_step_report"
        ]
      }
    },
...
  },
  "info": {
    "title": "ERP API",
    "description": "Dokumentasi API ERP multi-tenant",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "BaseQueryDtoSmart": {
        "type": "object",
        "properties": {
          "filter": {
            "type": "object",
            "description": "Where filter (supports >, >=, <, <=, like, notLike,  in, notIn, isNull, isNotNull)",
            "example": {
              "id": {
                "in": [
                  "id1",
                  "id2"
                ]
              },
              "status": {
                "eq": "aktif"
              },
              "created_at": {
                "between": [
                  "2024-01-01",
                  "2024-12-31"
                ]
              }
            },
            "additionalProperties": true
          },
          "include": {
            "type": "array",
            "description": "Nama tabel beserta tipenya single atau array",
            "example": [
              {
                "name": "contract",
                "type": "single",
                "select": [
                  "idContract",
                  "idClient",
                  "nomorKontrak",
                  "statusKontrak"
                ]
              },
              {
                "name": "client",
                "parent": "contract",
                "type": "single"
              },
              {
                "name": "work_schedule",
                "type": "array"
              }
            ],
            "additionalProperties": true
          },
          "joinWhere": {
            "type": "array",
            "description": "Filter + jenis JOIN (inner/left/right)",
            "example": [
              {
                "contract": {
                  "status_kontrak": "aktif"
                },
                "type": "inner"
              },
              {
                "client_site": {
                  "kota": "Jakarta"
                },
                "type": "left"
              }
            ]
          },
          "search_field": {
            "description": "Fields to search with search_keyword",
            "example": [
              "contract.nomor_kontrak",
              "site.kota"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "search_keyword": {
            "type": "string",
            "description": "Search keyword",
            "example": "PT"
          },
          "sortKey": {
            "type": "string",
            "description": "Sort by field",
            "example": "contract.nomor_kontrak"
          },
          "sortValue": {
            "type": "string",
            "description": "Sort direction (asc or desc)",
            "example": "asc"
          },
          "pageIndex": {
            "type": "number",
            "description": "Pagination page index",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "description": "Pagination limit per page",
            "example": 20
          }
        }
      },
...
      "ToolsWorkflowDto": {
        "type": "object",
        "properties": {
          "idWorkflow": {
            "type": "string"
          },
          "namaWorkflow": {
            "type": "string"
          },
          "forModule": {
            "type": "string"
          },
          "targetDb": {
            "type": "string"
          },
          "targetTable": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "idKantor": {
            "type": "string"
          },
          "keterangan": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "statusWorkflow": {
            "type": "string"
          }
        },
        "required": [
          "idWorkflow",
          "namaWorkflow",
          "forModule",
          "targetDb",
          "targetTable"
        ]
      },
...
      "ToolsWorkflowStepDto": {
        "type": "object",
        "properties": {
          "idWorkflowStep": {
            "type": "string"
          },
          "idWorkflow": {
            "type": "string"
          },
          "statusFrom": {
            "type": "string"
          },
          "statusTo": {
            "type": "string"
          },
          "urutan": {
            "type": "number"
          },
          "ruleJson": {
            "type": "string"
          },
          "triggerJson": {
            "type": "string"
          },
          "customCheckSql": {
            "type": "string"
          },
          "messageBefore": {
            "type": "string"
          },
          "statusRejectTo": {
            "type": "string"
          },
          "allowUserIds": {
            "type": "string"
          },
          "allowRoles": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "idWorkflowStep",
          "idWorkflow",
          "statusFrom",
          "statusTo",
          "urutan"
        ]
      },
      "ToolsWorkflowStepReportDto": {
        "type": "object",
        "properties": {
          "idWorkflowStep": {
            "type": "string"
          },
          "idWorkflow": {
            "type": "string"
          },
          "statusFrom": {
            "type": "string"
          },
          "statusTo": {
            "type": "string"
          },
          "urutan": {
            "type": "number"
          },
          "ruleJson": {
            "type": "string"
          },
          "triggerJson": {
            "type": "string"
          },
          "customCheckSql": {
            "type": "string"
          },
          "messageBefore": {
            "type": "string"
          },
          "statusRejectTo": {
            "type": "string"
          },
          "allowUserIds": {
            "type": "string"
          },
          "allowRoles": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "workflow": {
            "$ref": "#/components/schemas/ToolsWorkflowDto"
          }
        },
        "required": [
          "idWorkflowStep",
          "idWorkflow",
          "statusFrom",
          "statusTo",
          "urutan",
          "workflow"
        ]
      },
...
    }
  }
}